refactor mission component
