+++ src/pages/Landing/Landing.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import Header from 'pages/Landing/Header';
 import Highlight from 'pages/Landing/Hightlight';
 
-import Mission from '../../components/LandingPageComponents/Mission';
 import ComoFunciona from './ComoFunciona';
+import Mission from './Mission';
 
 const LandingPage = () => {
   return (
